react code refactoring: separation of components
