+++ react_native/components/openai.tsx
